Some refactoring and some structure changes.
