--- src/main/java/com/teachmeskills/lesson_21/task_2/model/User.java
@@ -1,46 +1,0 @@
-package com.teachmeskills.lesson_21.task_2.model;
-
-public class User {
-
-    private int id;
-    private String name;
-    private String email;
-
-    public User() { }
-
-    public User(String name, String email) {
-        this.name = name;
-        this.email = email;
-    }
-
-    public User(int id, String name, String email) {
-        this.id = id;
-        this.name = name;
-        this.email = email;
-    }
-
-    public int getId() {
-        return id;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-}
