+++ src/main/java/com/teachmeskills/lesson_21/task_2/servlet/AddUserServlet.java
@@ -1,4 +1,5 @@
 package com.teachmeskills.lesson_21.task_2.servlet;
 
 public class AddUserServlet {
+    // не нужен
 }
