+++ src/main/java/com/teachmeskills/lesson_21/task_2/servlet/GetListUserServlet.java
@@ -1,4 +1,5 @@
 package com.teachmeskills.lesson_21.task_2.servlet;
 
+// Не нужен
 public class GetListUserServlet {
 }
