+++ src/main/java/com/teachmeskills/lesson_21/task_2/servlet/UserServlet.java
@@ -1,8 +1,7 @@
 package com.teachmeskills.lesson_21.task_2.servlet;
 
-import com.teachmeskills.lesson_21.task_2.connector.MySQLConnector;
-import com.teachmeskills.lesson_21.task_2.crud.CRUD;
-import com.teachmeskills.lesson_21.task_2.model.User;
+import com.teachmeskills.lesson_21.task_2.service.UserCRUDService;
+import com.teachmeskills.lesson_21.task_2.entity.User;
 import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -11,7 +10,6 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.util.List;
-import java.util.concurrent.CopyOnWriteArrayList;
 
 @WebServlet("/")
 public class UserServlet extends HttpServlet {
@@ -19,21 +17,21 @@ public class UserServlet extends HttpServlet {
     private static final String INDEX_JSP = "/pages/index.jsp";
     private static final String UPDATE_JSP = "/pages/update.jsp";
     private List<User> listUsers;
-    private CRUD crud;
+    private UserCRUDService userCrudService;
     private String id;
     private String name;
     private String email;
 
     @Override
-    public void init() throws ServletException {
-        crud = new CRUD();
+    public void init() {
+        userCrudService = new UserCRUDService();
         //listUsers = new CopyOnWriteArrayList<>();
     }
 
     @Override
     public void destroy() {
         //MySQLConnector.closeThreadsClassMySQLConnector();
-        //crud = null;
+        //userCrudService = null;
         //listUsers = null;
         super.destroy();
     }
@@ -41,6 +39,7 @@ public class UserServlet extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         request.setAttribute("user", listUsers);
+        // TODO: Если пришел айди, то выводим одного. Если без айди, то выводим всех.
         request.getRequestDispatcher(INDEX_JSP).forward(request, response);
     }
 
@@ -48,6 +47,7 @@ public class UserServlet extends HttpServlet {
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         request.setCharacterEncoding("UTF8");
         String action = request.getServletPath();
+        // TODO: Здесь должно быть создание нового юзера с его параметрами. Все что ниже делают методы doDelete(), doPut().
 
         switch (action) {
             case "/add":                                   // добавление пользователя
@@ -75,18 +75,31 @@ public class UserServlet extends HttpServlet {
         doGet(request, response);
     }
 
-    private void newUser(HttpServletRequest req, HttpServletResponse resp) throws IOException, ServletException {
+    @Override
+    protected void doPut(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+        super.doPut(req, resp);
+        // TODO: руализовать модификацию Юзера по полученному айдишнику тут. Получаем параметры.
+        //  Айди - мандаторен. Все остальные параметры если есть, то обновляем для юзера по айди
+    }
+
+    @Override
+    protected void doDelete(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+        super.doDelete(req, resp);
+        // TODO: Удаляем юзера если пришел его айди. Если нет, то удаляем всех
+    }
+
+    private void newUser(HttpServletRequest req, HttpServletResponse resp) throws IOException {
         name = req.getParameter("name");
         email = req.getParameter("email");
         User newUser = new User(name, email);
-        crud.createUser(newUser);
+        userCrudService.createUser(newUser);
         resp.sendRedirect("list");
         //doGet(req,resp);
     }
 
     private void getDataById(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         id = req.getParameter("id");
-        User user = crud.getUserById(Integer.parseInt(id));
+        User user = userCrudService.getUserById(Integer.parseInt(id));
         RequestDispatcher dispatcher = req.getRequestDispatcher(INDEX_JSP);
         req.setAttribute("user", user);
         dispatcher.forward(req, resp);
@@ -102,7 +115,7 @@ public class UserServlet extends HttpServlet {
         name = req.getParameter("name");
         email = req.getParameter("email");
         User user = new User(Integer.parseInt(id), name, email);
-        crud.updateUser(user);
+        userCrudService.updateUser(user);
         req.setAttribute("user", user);
         req.getRequestDispatcher(UPDATE_JSP).forward(req, resp);
         resp.sendRedirect("list");
@@ -110,17 +123,17 @@ public class UserServlet extends HttpServlet {
 
     private void deleteById(HttpServletRequest req, HttpServletResponse resp) throws IOException {
         id = req.getParameter("id");
-        crud.deleteUserById(Integer.parseInt(id));
+        userCrudService.deleteUserById(Integer.parseInt(id));
         resp.sendRedirect("list");
     }
 
     private void deleteAll(HttpServletResponse resp) throws IOException {
-        crud.deleteAllUsers();
+        userCrudService.deleteAllUsers();
         resp.sendRedirect("list");
     }
 
     private void listUsers(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        listUsers = crud.getDataAllUsers();
+        listUsers = userCrudService.getDataAllUsers();
         req.setAttribute("listUser", listUsers);
         RequestDispatcher dispatcher = req.getRequestDispatcher(INDEX_JSP);
         dispatcher.forward(req, resp);
